pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openclaw-memory"
version = "0.1.0"
description = "Three-layer memory system for chatbot applications"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "psycopg[binary]>=3.1",
    "psycopg-pool>=3.1",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
gemini = ["google-generativeai>=0.8"]
voyage = ["voyageai>=0.3"]
local = ["llama-cpp-python>=0.2"]
all = [
    "openai>=1.0",
    "anthropic>=0.30",
    "google-generativeai>=0.8",
    "voyageai>=0.3",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "ruff>=0.4",
    "mypy>=1.10",
    "openai>=1.0",
    "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/openclaw_memory"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-v", "--tb=short"]
